# openapi.yaml
openapi: 3.1.1
info:
  title: CarConfigurator API
  version: 1.0.0
  description: >-
    Comprehensive car configuration API with internationalization, user management, and admin features. Fully compliant
    with OpenAPI 3.1.1 specification for Robot Framework training and automation testing.
  contact:
    name: CarConfigurator Team
    email: support@carconfigurator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
components:
  schemas:
    Car:
      type: object
      required:
        - id
        - name
        - category
        - basePrice
        - description
      properties:
        id:
          type: string
          description: Unique car identifier
        name:
          type: string
          description: Car model name
          examples:
            - BMW X5
            - Audi Q7
            - Mercedes GLE
        category:
          type: string
          description: Car category
          examples:
            - SUV
            - Sedan
            - Coupe
        basePrice:
          type: number
          format: float
          description: Base price in Euro
          examples:
            - 45000
            - 52000
            - 68000
        description:
          type: string
          description: Car description
          examples:
            - Luxury SUV with latest technology
            - Premium sedan with hybrid drive
        imageUrl:
          type:
            - string
            - 'null'
          description: Car image URL (can be relative or absolute)
          examples:
            - /images/cars/elegance-sedan.jpg
            - https://example.com/car.jpg
        imageData:
          type:
            - string
            - 'null'
          description: Base64 encoded image data
        imageMimeType:
          type:
            - string
            - 'null'
          description: Image MIME type
          examples:
            - image/jpeg
            - image/png
            - image/webp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
        data:
          description: 'Response data (only when success: true)'
        error:
          type: string
          description: 'Error message (only when success: false)'
...
    LocaleUpdate:
      type: object
      required:
        - locale
      properties:
        locale:
          type: string
          enum:
            - de
            - en
          description: Language locale code
...
  responses:
    ValidationError:
      description: Validation error response
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
    InternalServerError:
      description: Internal server error response
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
...
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: cookie
      name: admin-session
      description: Admin-Session-Cookie für authentifizierte Anfragen
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-Token für Benutzerauthentifizierung
paths:
...
  /api/auth/me:
    get:
      summary: Get current authenticated user
      description: Returns the current user's data if the provided token is valid
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: user_123
                      email:
                        type: string
                        example: user@example.com
                      name:
                        type: string
                        example: John Doe
                      role:
                        type: string
                        example: user
        '401':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: Token ungültig
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: Server-Fehler
...
  /api/cars/{id}:
    get:
      tags:
        - Cars
      summary: Get car by ID
      description: Retrieves a specific car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Car ID
        - name: locale
          in: query
          schema:
            type: string
            enum:
              - de
              - en
            default: de
          description: Language for localized content
      responses:
        '200':
          description: Car successfully retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Car'
        '404':
          description: Car not found
        '500':
          description: Internal server error
    put:
      tags:
        - Cars
      summary: Update car
      description: Updates an existing car (admin only)
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Car ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  description: Car model name
                category:
                  type: string
                  minLength: 1
                  description: Car category
                basePrice:
                  type: number
                  format: float
                  minimum: 0.01
                  description: Base price in Euro
                description:
                  type: string
                  minLength: 1
                  description: Car description
                imageUrl:
                  type: string
                  format: uri
                  description: Car image URL
                imageData:
                  type: string
                  description: Base64 encoded image data
                imageMimeType:
                  type: string
                  description: Image MIME type
                translations:
                  type: array
                  items:
                    type: object
                    properties:
                      locale:
                        type: string
                        minLength: 2
                        examples:
                          - de
                          - en
                      name:
                        type: string
                        minLength: 1
                      category:
                        type: string
                        minLength: 1
                      description:
                        type: string
                        minLength: 1
      responses:
        '200':
          description: Car successfully updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Car'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Car not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Cars
      summary: Delete car
      description: Deletes a car (admin only)
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Car ID
      responses:
        '200':
          description: Car successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '404':
          description: Car not found
        '500':
          description: Internal server error
...
  /api/locale:
    post:
      tags:
        - Locale
      summary: Update user locale
      description: Updates the user's preferred language locale (de or en)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocaleUpdate'
      responses:
        '200':
          description: Locale updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...